[Refactor] Move statements from ir.h to statements.h
