+++ taichi/ir/ir.cpp
@@ -318,223 +318,6 @@ std::string ExprGroup::serialize() const {
   return ret;
 }
 
-UnaryOpStmt::UnaryOpStmt(UnaryOpType op_type, Stmt *operand)
-    : op_type(op_type), operand(operand) {
-  TI_ASSERT(!operand->is<AllocaStmt>());
-  cast_type = DataType::unknown;
-  TI_STMT_REG_FIELDS;
-}
-
-bool UnaryOpStmt::is_cast() const {
-  return unary_op_is_cast(op_type);
-}
-
-bool UnaryOpStmt::same_operation(UnaryOpStmt *o) const {
-  if (op_type == o->op_type) {
-    if (is_cast()) {
-      return cast_type == o->cast_type;
-    } else {
-      return true;
-    }
-  }
-  return false;
-}
-
-std::string UnaryOpExpression::serialize() {
-  if (is_cast()) {
-    std::string reint = type == UnaryOpType::cast_value ? "" : "reinterpret_";
-    return fmt::format("({}{}<{}> {})", reint, unary_op_type_name(type),
-                       data_type_name(cast_type), operand->serialize());
-  } else {
-    return fmt::format("({} {})", unary_op_type_name(type),
-                       operand->serialize());
-  }
-}
-
-bool UnaryOpExpression::is_cast() const {
-  return unary_op_is_cast(type);
-}
-
-void UnaryOpExpression::flatten(FlattenContext *ctx) {
-  operand->flatten(ctx);
-  auto unary = std::make_unique<UnaryOpStmt>(type, operand->stmt);
-  if (is_cast()) {
-    unary->cast_type = cast_type;
-  }
-  stmt = unary.get();
-  stmt->tb = tb;
-  ctx->push_back(std::move(unary));
-}
-
-void ExternalFuncCallExpression::flatten(FlattenContext *ctx) {
-  std::vector<Stmt *> arg_statements, output_statements;
-  for (auto &s : args) {
-    s->flatten(ctx);
-    arg_statements.push_back(s->stmt);
-  }
-  for (auto &s : outputs) {
-    output_statements.push_back(s.cast<IdExpression>()->flatten_noload(ctx));
-  }
-  ctx->push_back(std::make_unique<ExternalFuncCallStmt>(
-      func, source, arg_statements, output_statements));
-  stmt = ctx->back_stmt();
-}
-
-ExternalPtrStmt::ExternalPtrStmt(const LaneAttribute<Stmt *> &base_ptrs,
-                                 const std::vector<Stmt *> &indices)
-    : base_ptrs(base_ptrs), indices(indices) {
-  DataType dt = DataType::f32;
-  for (int i = 0; i < (int)base_ptrs.size(); i++) {
-    TI_ASSERT(base_ptrs[i] != nullptr);
-    TI_ASSERT(base_ptrs[i]->is<ArgLoadStmt>());
-  }
-  width() = base_ptrs.size();
-  element_type() = dt;
-  TI_STMT_REG_FIELDS;
-}
-
-GlobalPtrStmt::GlobalPtrStmt(const LaneAttribute<SNode *> &snodes,
-                             const std::vector<Stmt *> &indices,
-                             bool activate)
-    : snodes(snodes), indices(indices), activate(activate) {
-  for (int i = 0; i < (int)snodes.size(); i++) {
-    TI_ASSERT(snodes[i] != nullptr);
-    TI_ASSERT(snodes[0]->dt == snodes[i]->dt);
-  }
-  width() = snodes.size();
-  element_type() = snodes[0]->dt;
-  TI_STMT_REG_FIELDS;
-}
-
-bool GlobalPtrStmt::is_element_wise(SNode *snode) const {
-  if (snode == nullptr) {
-    // check every SNode when "snode" is nullptr
-    for (const auto &snode_i : snodes.data) {
-      if (!is_element_wise(snode_i)) {
-        return false;
-      }
-    }
-    return true;
-  }
-  // check if this statement is element-wise on a specific SNode, i.e., argument
-  // "snode"
-  for (int i = 0; i < (int)indices.size(); i++) {
-    if (auto loop_index_i = indices[i]->cast<LoopIndexStmt>();
-        !(loop_index_i && loop_index_i->loop->is<OffloadedStmt>() &&
-          loop_index_i->index == snode->physical_index_position[i])) {
-      return false;
-    }
-  }
-  return true;
-}
-
-std::string GlobalPtrExpression::serialize() {
-  std::string s = fmt::format("{}[", var.serialize());
-  for (int i = 0; i < (int)indices.size(); i++) {
-    s += indices.exprs[i]->serialize();
-    if (i + 1 < (int)indices.size())
-      s += ", ";
-  }
-  s += "]";
-  return s;
-}
-
-void GlobalPtrExpression::flatten(FlattenContext *ctx) {
-  std::vector<Stmt *> index_stmts;
-  for (int i = 0; i < (int)indices.size(); i++) {
-    indices.exprs[i]->flatten(ctx);
-    index_stmts.push_back(indices.exprs[i]->stmt);
-  }
-  if (var.is<GlobalVariableExpression>()) {
-    ctx->push_back(std::make_unique<GlobalPtrStmt>(
-        var.cast<GlobalVariableExpression>()->snode, index_stmts));
-  } else {
-    TI_ASSERT(var.is<ExternalTensorExpression>());
-    var->flatten(ctx);
-    ctx->push_back(std::make_unique<ExternalPtrStmt>(
-        var.cast<ExternalTensorExpression>()->stmt, index_stmts));
-  }
-  stmt = ctx->back_stmt();
-}
-
-GetChStmt::GetChStmt(Stmt *input_ptr, int chid)
-    : input_ptr(input_ptr), chid(chid) {
-  TI_ASSERT(input_ptr->is<SNodeLookupStmt>());
-  input_snode = input_ptr->as<SNodeLookupStmt>()->snode;
-  output_snode = input_snode->ch[chid].get();
-  TI_STMT_REG_FIELDS;
-}
-
-DecoratorRecorder dec;
-
-FrontendContext::FrontendContext() {
-  root_node = std::make_unique<Block>();
-  current_builder = std::make_unique<IRBuilder>(root_node.get());
-}
-
-FrontendForStmt::FrontendForStmt(const Expr &loop_var,
-                                 const Expr &begin,
-                                 const Expr &end)
-    : begin(begin), end(end) {
-  vectorize = dec.vectorize;
-  parallelize = dec.parallelize;
-  strictly_serialized = dec.strictly_serialized;
-  block_dim = dec.block_dim;
-  auto cfg = get_current_program().config;
-  if (cfg.arch == Arch::cuda) {
-    vectorize = 1;
-    parallelize = 1;
-  } else {
-    if (parallelize == 0)
-      parallelize = std::thread::hardware_concurrency();
-  }
-  scratch_opt = dec.scratch_opt;
-  dec.reset();
-  if (vectorize == -1)
-    vectorize = 1;
-  loop_var_id.resize(1);
-  loop_var_id[0] = loop_var.cast<IdExpression>()->id;
-}
-
-FrontendForStmt::FrontendForStmt(const ExprGroup &loop_var,
-                                 const Expr &global_var)
-    : global_var(global_var) {
-  vectorize = dec.vectorize;
-  parallelize = dec.parallelize;
-  strictly_serialized = dec.strictly_serialized;
-  block_dim = dec.block_dim;
-  auto cfg = get_current_program().config;
-  if (cfg.arch == Arch::cuda) {
-    vectorize = 1;
-    parallelize = 1;
-    TI_ASSERT(block_dim <= taichi_max_gpu_block_dim);
-  } else {
-    // cpu
-    if (parallelize == 0)
-      parallelize = std::thread::hardware_concurrency();
-  }
-  scratch_opt = dec.scratch_opt;
-  dec.reset();
-  if (vectorize == -1)
-    vectorize = 1;
-
-  loop_var_id.resize(loop_var.size());
-  for (int i = 0; i < (int)loop_var.size(); i++) {
-    loop_var_id[i] = loop_var[i].cast<IdExpression>()->id;
-  }
-}
-
-FrontendAssignStmt::FrontendAssignStmt(const Expr &lhs, const Expr &rhs)
-    : lhs(lhs), rhs(rhs) {
-  TI_ASSERT(lhs->is_lvalue());
-}
-
-IRNode *FrontendContext::root() {
-  return static_cast<IRNode *>(root_node.get());
-}
-
-std::unique_ptr<FrontendContext> context;
-
 template <>
 std::string to_string(const LaneAttribute<LocalAddress> &ptr) {
   std::string ret = " [";
@@ -545,74 +328,6 @@ std::string to_string(const LaneAttribute<LocalAddress> &ptr) {
   }
   ret += "]";
   return ret;
-}
-
-Stmt *LocalLoadStmt::previous_store_or_alloca_in_block() {
-  int position = parent->locate(this);
-  // TI_ASSERT(width() == 1);
-  // TI_ASSERT(this->ptr[0].offset == 0);
-  for (int i = position - 1; i >= 0; i--) {
-    if (parent->statements[i]->is<LocalStoreStmt>()) {
-      auto store = parent->statements[i]->as<LocalStoreStmt>();
-      // TI_ASSERT(store->width() == 1);
-      if (store->ptr == this->ptr[0].var) {
-        // found
-        return store;
-      }
-    } else if (parent->statements[i]->is<AllocaStmt>()) {
-      auto alloca = parent->statements[i]->as<AllocaStmt>();
-      // TI_ASSERT(alloca->width() == 1);
-      if (alloca == this->ptr[0].var) {
-        return alloca;
-      }
-    }
-  }
-  return nullptr;
-}
-
-bool LocalLoadStmt::same_source() const {
-  for (int i = 1; i < (int)ptr.size(); i++) {
-    if (ptr[i].var != ptr[0].var)
-      return false;
-  }
-  return true;
-}
-
-bool LocalLoadStmt::has_source(Stmt *alloca) const {
-  for (int i = 0; i < width(); i++) {
-    if (ptr[i].var == alloca)
-      return true;
-  }
-  return false;
-}
-
-IfStmt::IfStmt(Stmt *cond)
-    : cond(cond), true_mask(nullptr), false_mask(nullptr) {
-  TI_STMT_REG_FIELDS;
-}
-
-void IfStmt::set_true_statements(std::unique_ptr<Block> &&new_true_statements) {
-  true_statements = std::move(new_true_statements);
-  if (true_statements)
-    true_statements->parent_stmt = this;
-}
-
-void IfStmt::set_false_statements(
-    std::unique_ptr<Block> &&new_false_statements) {
-  false_statements = std::move(new_false_statements);
-  if (false_statements)
-    false_statements->parent_stmt = this;
-}
-
-std::unique_ptr<Stmt> IfStmt::clone() const {
-  auto new_stmt = std::make_unique<IfStmt>(cond);
-  new_stmt->true_mask = true_mask;
-  new_stmt->false_mask = false_mask;
-  if (true_statements)
-    new_stmt->set_true_statements(true_statements->clone());
-  if (false_statements)
-    new_stmt->set_false_statements(false_statements->clone());
-  return new_stmt;
 }
 
 void Block::erase(int location) {
@@ -876,242 +591,12 @@ bool DelayedIRModifier::modify_ir() {
   return true;
 }
 
-FrontendSNodeOpStmt::FrontendSNodeOpStmt(SNodeOpType op_type,
-                                         SNode *snode,
-                                         const ExprGroup &indices,
-                                         const Expr &val)
-    : op_type(op_type), snode(snode), indices(indices.loaded()), val(val) {
-  if (val.expr != nullptr) {
-    TI_ASSERT(op_type == SNodeOpType::append);
-    this->val.set(load_if_ptr(val));
-  } else {
-    TI_ASSERT(op_type != SNodeOpType::append);
-  }
-}
-
-SNodeOpStmt::SNodeOpStmt(SNodeOpType op_type,
-                         SNode *snode,
-                         Stmt *ptr,
-                         Stmt *val)
-    : op_type(op_type), snode(snode), ptr(ptr), val(val) {
-  width() = 1;
-  element_type() = DataType::i32;
-  TI_STMT_REG_FIELDS;
-}
-
-SNodeOpStmt::SNodeOpStmt(SNodeOpType op_type,
-                         SNode *snode,
-                         const std::vector<Stmt *> &indices)
-    : op_type(op_type), snode(snode), indices(indices) {
-  ptr = nullptr;
-  val = nullptr;
-  TI_ASSERT(op_type == SNodeOpType::is_active ||
-            op_type == SNodeOpType::deactivate ||
-            op_type == SNodeOpType::activate);
-  width() = 1;
-  element_type() = DataType::i32;
-  TI_STMT_REG_FIELDS;
-}
-
-bool SNodeOpStmt::activation_related(SNodeOpType op) {
-  return op == SNodeOpType::activate || op == SNodeOpType::deactivate ||
-         op == SNodeOpType::is_active;
-}
-
-bool SNodeOpStmt::need_activation(SNodeOpType op) {
-  return op == SNodeOpType::activate || op == SNodeOpType::append;
-}
-
-std::string AtomicOpExpression::serialize() {
-  if (op_type == AtomicOpType::add) {
-    return fmt::format("atomic_add({}, {})", dest.serialize(), val.serialize());
-  } else if (op_type == AtomicOpType::sub) {
-    return fmt::format("atomic_sub({}, {})", dest.serialize(), val.serialize());
-  } else if (op_type == AtomicOpType::min) {
-    return fmt::format("atomic_min({}, {})", dest.serialize(), val.serialize());
-  } else if (op_type == AtomicOpType::max) {
-    return fmt::format("atomic_max({}, {})", dest.serialize(), val.serialize());
-  } else if (op_type == AtomicOpType::bit_and) {
-    return fmt::format("atomic_bit_and({}, {})", dest.serialize(),
-                       val.serialize());
-  } else if (op_type == AtomicOpType::bit_or) {
-    return fmt::format("atomic_bit_or({}, {})", dest.serialize(),
-                       val.serialize());
-  } else if (op_type == AtomicOpType::bit_xor) {
-    return fmt::format("atomic_bit_xor({}, {})", dest.serialize(),
-                       val.serialize());
-  } else {
-    // min/max not supported in the LLVM backend yet.
-    TI_NOT_IMPLEMENTED;
-  }
-}
-
-void AtomicOpExpression::flatten(FlattenContext *ctx) {
-  // replace atomic sub with negative atomic add
-  if (op_type == AtomicOpType::sub) {
-    val.set(Expr::make<UnaryOpExpression>(UnaryOpType::neg, val));
-    op_type = AtomicOpType::add;
-  }
-  // expand rhs
-  auto expr = val;
-  expr->flatten(ctx);
-  if (dest.is<IdExpression>()) {  // local variable
-    // emit local store stmt
-    auto alloca = ctx->current_block->lookup_var(dest.cast<IdExpression>()->id);
-    ctx->push_back<AtomicOpStmt>(op_type, alloca, expr->stmt);
-  } else {  // global variable
-    TI_ASSERT(dest.is<GlobalPtrExpression>());
-    auto global_ptr = dest.cast<GlobalPtrExpression>();
-    global_ptr->flatten(ctx);
-    ctx->push_back<AtomicOpStmt>(op_type, ctx->back_stmt(), expr->stmt);
-  }
-  stmt = ctx->back_stmt();
-}
-
-std::string SNodeOpExpression::serialize() {
-  if (value.expr) {
-    return fmt::format("{}({}, [{}], {})", snode_op_type_name(op_type),
-                       snode->get_node_type_name_hinted(), indices.serialize(),
-                       value.serialize());
-  } else {
-    return fmt::format("{}({}, [{}])", snode_op_type_name(op_type),
-                       snode->get_node_type_name_hinted(), indices.serialize());
-  }
-}
-
-void SNodeOpExpression::flatten(FlattenContext *ctx) {
-  std::vector<Stmt *> indices_stmt;
-  for (int i = 0; i < (int)indices.size(); i++) {
-    indices[i]->flatten(ctx);
-    indices_stmt.push_back(indices[i]->stmt);
-  }
-  if (op_type == SNodeOpType::is_active) {
-    // is_active cannot be lowered all the way to a global pointer.
-    // It should be lowered into a pointer to parent and an index.
-    TI_ERROR_IF(snode->type != SNodeType::pointer &&
-                    snode->type != SNodeType::hash &&
-                    snode->type != SNodeType::bitmasked,
-                "ti.is_active only works on pointer, hash or bitmasked nodes.");
-    ctx->push_back<SNodeOpStmt>(SNodeOpType::is_active, snode, indices_stmt);
-  } else {
-    auto ptr = ctx->push_back<GlobalPtrStmt>(snode, indices_stmt);
-    if (op_type == SNodeOpType::append) {
-      value->flatten(ctx);
-      ctx->push_back<SNodeOpStmt>(SNodeOpType::append, snode, ptr, value->stmt);
-      TI_ERROR_IF(snode->type != SNodeType::dynamic,
-                  "ti.append only works on dynamic nodes.");
-      TI_ERROR_IF(snode->ch.size() != 1,
-                  "ti.append only works on single-child dynamic nodes.");
-      TI_ERROR_IF(data_type_size(snode->ch[0]->dt) != 4,
-                  "ti.append only works on i32/f32 nodes.");
-    } else if (op_type == SNodeOpType::length) {
-      ctx->push_back<SNodeOpStmt>(SNodeOpType::length, snode, ptr, nullptr);
-    }
-  }
-  stmt = ctx->back_stmt();
-}
-
-std::unique_ptr<ConstStmt> ConstStmt::copy() {
-  return std::make_unique<ConstStmt>(val);
-}
-
-RangeForStmt::RangeForStmt(Stmt *begin,
-                           Stmt *end,
-                           std::unique_ptr<Block> &&body,
-                           int vectorize,
-                           int parallelize,
-                           int block_dim,
-                           bool strictly_serialized)
-    : begin(begin),
-      end(end),
-      body(std::move(body)),
-      vectorize(vectorize),
-      parallelize(parallelize),
-      block_dim(block_dim),
-      strictly_serialized(strictly_serialized) {
-  reversed = false;
-  this->body->parent_stmt = this;
-  TI_STMT_REG_FIELDS;
-}
-
-std::unique_ptr<Stmt> RangeForStmt::clone() const {
-  auto new_stmt = std::make_unique<RangeForStmt>(
-      begin, end, body->clone(), vectorize, parallelize, block_dim,
-      strictly_serialized);
-  new_stmt->reversed = reversed;
-  return new_stmt;
-}
-
-StructForStmt::StructForStmt(SNode *snode,
-                             std::unique_ptr<Block> &&body,
-                             int vectorize,
-                             int parallelize,
-                             int block_dim)
-    : snode(snode),
-      body(std::move(body)),
-      vectorize(vectorize),
-      parallelize(parallelize),
-      block_dim(block_dim) {
-  this->body->parent_stmt = this;
-  TI_STMT_REG_FIELDS;
-}
-
-std::unique_ptr<Stmt> StructForStmt::clone() const {
-  auto new_stmt = std::make_unique<StructForStmt>(
-      snode, body->clone(), vectorize, parallelize, block_dim);
-  new_stmt->scratch_opt = scratch_opt;
-  return new_stmt;
-}
-
-FuncBodyStmt::FuncBodyStmt(const std::string &funcid,
-                           std::unique_ptr<Block> &&body)
-    : funcid(funcid), body(std::move(body)) {
-  if (this->body)
-    this->body->parent_stmt = this;
-  TI_STMT_REG_FIELDS;
-}
-
-std::unique_ptr<Stmt> FuncBodyStmt::clone() const {
-  return std::make_unique<FuncBodyStmt>(funcid, body->clone());
-}
-
-WhileStmt::WhileStmt(std::unique_ptr<Block> &&body)
-    : mask(nullptr), body(std::move(body)) {
-  this->body->parent_stmt = this;
-  TI_STMT_REG_FIELDS;
-}
-
-std::unique_ptr<Stmt> WhileStmt::clone() const {
-  auto new_stmt = std::make_unique<WhileStmt>(body->clone());
-  new_stmt->mask = mask;
-  return new_stmt;
-}
-
-bool ContinueStmt::as_return() const {
-  TI_ASSERT(scope != nullptr);
-  if (auto *offl = scope->cast<OffloadedStmt>(); offl) {
-    TI_ASSERT(offl->task_type == OffloadedStmt::TaskType::range_for ||
-              offl->task_type == OffloadedStmt::TaskType::struct_for);
-    return true;
-  }
-  return false;
+LocalAddress::LocalAddress(Stmt *var, int offset) : var(var), offset(offset) {
+  TI_ASSERT(var->is<AllocaStmt>());
 }
 
 void Stmt::infer_type() {
   irpass::type_check(this);
-}
-
-ExternalTensorShapeAlongAxisStmt::ExternalTensorShapeAlongAxisStmt(int axis,
-                                                                   int arg_id)
-    : axis(axis), arg_id(arg_id) {
-  TI_STMT_REG_FIELDS;
-}
-
-void ExternalTensorShapeAlongAxisExpression::flatten(FlattenContext *ctx) {
-  auto temp = ptr.cast<ExternalTensorExpression>();
-  TI_ASSERT(0 <= axis && axis < temp->dim);
-  ctx->push_back<ExternalTensorShapeAlongAxisStmt>(axis, temp->arg_id);
-  stmt = ctx->back_stmt();
 }
 
 TLANG_NAMESPACE_END
