+++ taichi/ir/statements.h
@@ -5,6 +5,639 @@
 
 TLANG_NAMESPACE_BEGIN
 
+class AllocaStmt : public Stmt {
+ public:
+  AllocaStmt(DataType type) {
+    ret_type = VectorType(1, type);
+    TI_STMT_REG_FIELDS;
+  }
+
+  AllocaStmt(int width, DataType type) {
+    ret_type = VectorType(width, type);
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+// updates mask, break if no active
+class WhileControlStmt : public Stmt {
+ public:
+  Stmt *mask;
+  Stmt *cond;
+  WhileControlStmt(Stmt *mask, Stmt *cond) : mask(mask), cond(cond) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(mask, cond);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class ContinueStmt : public Stmt {
+ public:
+  // This is the loop on which this continue stmt has effects. It can be either
+  // an offloaded task, or a for/while loop inside the kernel.
+  Stmt *scope;
+
+  ContinueStmt() : scope(nullptr) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  // For top-level loops, since they are parallelized to multiple threads (on
+  // either CPU or GPU), `continue` becomes semantically equivalent to `return`.
+  //
+  // Caveat:
+  // We should wrap each backend's kernel body into a function (as LLVM does).
+  // The reason is that, each thread may handle more than one element,
+  // depending on the backend's implementation.
+  //
+  // For example, CUDA uses gride-stride loops, the snippet below illustrates
+  // the idea:
+  //
+  // __global__ foo_kernel(...) {
+  //   for (int i = lower; i < upper; i += gridDim) {
+  //     auto coord = compute_coords(i);
+  //     // run_foo_kernel is produced by codegen
+  //     run_foo_kernel(coord);
+  //   }
+  // }
+  //
+  // If run_foo_kernel() is directly inlined within foo_kernel(), `return`
+  // could prematurely terminate the entire kernel.
+  bool as_return() const;
+
+  TI_STMT_DEF_FIELDS(scope);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class UnaryOpStmt : public Stmt {
+ public:
+  UnaryOpType op_type;
+  Stmt *operand;
+  DataType cast_type;
+
+  UnaryOpStmt(UnaryOpType op_type, Stmt *operand);
+
+  bool same_operation(UnaryOpStmt *o) const;
+  bool is_cast() const;
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, op_type, operand, cast_type);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class ArgLoadStmt : public Stmt {
+ public:
+  int arg_id;
+
+  ArgLoadStmt(int arg_id, DataType dt, bool is_ptr = false) : arg_id(arg_id) {
+    this->ret_type = VectorType(1, dt);
+    this->is_ptr = is_ptr;
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, arg_id, is_ptr);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class RandStmt : public Stmt {
+ public:
+  RandStmt(DataType dt) {
+    ret_type.data_type = dt;
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class BinaryOpStmt : public Stmt {
+ public:
+  BinaryOpType op_type;
+  Stmt *lhs, *rhs;
+
+  BinaryOpStmt(BinaryOpType op_type, Stmt *lhs, Stmt *rhs)
+      : op_type(op_type), lhs(lhs), rhs(rhs) {
+    TI_ASSERT(!lhs->is<AllocaStmt>());
+    TI_ASSERT(!rhs->is<AllocaStmt>());
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, op_type, lhs, rhs);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class TernaryOpStmt : public Stmt {
+ public:
+  TernaryOpType op_type;
+  Stmt *op1, *op2, *op3;
+
+  TernaryOpStmt(TernaryOpType op_type, Stmt *op1, Stmt *op2, Stmt *op3)
+      : op_type(op_type), op1(op1), op2(op2), op3(op3) {
+    TI_ASSERT(!op1->is<AllocaStmt>());
+    TI_ASSERT(!op2->is<AllocaStmt>());
+    TI_ASSERT(!op3->is<AllocaStmt>());
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, op1, op2, op3);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class AtomicOpStmt : public Stmt {
+ public:
+  AtomicOpType op_type;
+  Stmt *dest, *val;
+
+  AtomicOpStmt(AtomicOpType op_type, Stmt *dest, Stmt *val)
+      : op_type(op_type), dest(dest), val(val) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, op_type, dest, val);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class ExternalPtrStmt : public Stmt {
+ public:
+  LaneAttribute<Stmt *> base_ptrs;
+  std::vector<Stmt *> indices;
+
+  ExternalPtrStmt(const LaneAttribute<Stmt *> &base_ptrs,
+                  const std::vector<Stmt *> &indices);
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, base_ptrs, indices);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class GlobalPtrStmt : public Stmt {
+ public:
+  LaneAttribute<SNode *> snodes;
+  std::vector<Stmt *> indices;
+  bool activate;
+
+  GlobalPtrStmt(const LaneAttribute<SNode *> &snodes,
+                const std::vector<Stmt *> &indices,
+                bool activate = true);
+
+  bool is_element_wise(SNode *snode) const;
+
+  bool has_global_side_effect() const override {
+    return activate;
+  }
+
+  bool common_statement_eliminable() const override {
+    return true;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, snodes, indices, activate);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class SNodeOpStmt : public Stmt {
+ public:
+  SNodeOpType op_type;
+  SNode *snode;
+  Stmt *ptr;
+  Stmt *val;
+  std::vector<Stmt *> indices;
+
+  SNodeOpStmt(SNodeOpType op_type,
+              SNode *snode,
+              Stmt *ptr,
+              Stmt *val = nullptr);
+
+  SNodeOpStmt(SNodeOpType op_type,
+              SNode *snode,
+              const std::vector<Stmt *> &indices);
+
+  static bool activation_related(SNodeOpType op);
+
+  static bool need_activation(SNodeOpType op);
+
+  TI_STMT_DEF_FIELDS(ret_type, op_type, snode, ptr, val, indices);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class ExternalTensorShapeAlongAxisStmt : public Stmt {
+ public:
+  int axis;
+  int arg_id;
+
+  ExternalTensorShapeAlongAxisStmt(int axis, int arg_id);
+
+  TI_STMT_DEF_FIELDS(ret_type, axis, arg_id);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class AssertStmt : public Stmt {
+ public:
+  Stmt *cond;
+  std::string text;
+  std::vector<Stmt *> args;
+
+  AssertStmt(Stmt *cond,
+             const std::string &text,
+             const std::vector<Stmt *> &args)
+      : cond(cond), text(text), args(args) {
+    TI_ASSERT(cond);
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(cond, text, args);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class ExternalFuncCallStmt : public Stmt {
+ public:
+  void *func;
+  std::string source;
+  std::vector<Stmt *> arg_stmts;
+  std::vector<Stmt *> output_stmts;
+
+  ExternalFuncCallStmt(void *func,
+                       std::string const &source,
+                       const std::vector<Stmt *> &arg_stmts,
+                       const std::vector<Stmt *> &output_stmts)
+      : func(func),
+        source(source),
+        arg_stmts(arg_stmts),
+        output_stmts(output_stmts) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(func, arg_stmts, output_stmts);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class RangeAssumptionStmt : public Stmt {
+ public:
+  Stmt *input;
+  Stmt *base;
+  int low, high;
+
+  RangeAssumptionStmt(Stmt *input, Stmt *base, int low, int high)
+      : input(input), base(base), low(low), high(high) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, input, base, low, high);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class GlobalLoadStmt : public Stmt {
+ public:
+  Stmt *ptr;
+
+  GlobalLoadStmt(Stmt *ptr) : ptr(ptr) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, ptr);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class GlobalStoreStmt : public Stmt {
+ public:
+  Stmt *ptr, *data;
+
+  GlobalStoreStmt(Stmt *ptr, Stmt *data) : ptr(ptr), data(data) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, ptr, data);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class LocalLoadStmt : public Stmt {
+ public:
+  LaneAttribute<LocalAddress> ptr;
+
+  LocalLoadStmt(const LaneAttribute<LocalAddress> &ptr) : ptr(ptr) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool same_source() const;
+  bool has_source(Stmt *alloca) const;
+
+  Stmt *previous_store_or_alloca_in_block();
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, ptr);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class LocalStoreStmt : public Stmt {
+ public:
+  Stmt *ptr;
+  Stmt *data;
+
+  LocalStoreStmt(Stmt *ptr, Stmt *data) : ptr(ptr), data(data) {
+    TI_ASSERT(ptr->is<AllocaStmt>());
+    TI_STMT_REG_FIELDS;
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  bool dead_instruction_eliminable() const override {
+    return false;
+  }
+
+  bool common_statement_eliminable() const override {
+    return false;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, ptr, data);
+  TI_DEFINE_ACCEPT_AND_CLONE;
+};
+
+class IfStmt : public Stmt {
+ public:
+  Stmt *cond;
+  Stmt *true_mask, *false_mask;
+  std::unique_ptr<Block> true_statements, false_statements;
+
+  IfStmt(Stmt *cond);
+
+  // Use these setters to set Block::parent_stmt at the same time.
+  void set_true_statements(std::unique_ptr<Block> &&new_true_statements);
+  void set_false_statements(std::unique_ptr<Block> &&new_false_statements);
+
+  bool is_container_statement() const override {
+    return true;
+  }
+
+  std::unique_ptr<Stmt> clone() const override;
+
+  TI_STMT_DEF_FIELDS(cond, true_mask, false_mask);
+  TI_DEFINE_ACCEPT
+};
+
+class PrintStmt : public Stmt {
+ public:
+  using EntryType = std::variant<Stmt *, std::string>;
+  std::vector<EntryType> contents;
+
+  PrintStmt(const std::vector<EntryType> &contents_) : contents(contents_) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  template <typename... Args>
+  PrintStmt(Stmt *t, Args &&... args)
+      : contents(make_entries(t, std::forward<Args>(args)...)) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  template <typename... Args>
+  PrintStmt(const std::string &str, Args &&... args)
+      : contents(make_entries(str, std::forward<Args>(args)...)) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, contents);
+  TI_DEFINE_ACCEPT_AND_CLONE
+
+ private:
+  static void make_entries_helper(std::vector<PrintStmt::EntryType> &entries) {
+  }
+
+  template <typename T, typename... Args>
+  static void make_entries_helper(std::vector<PrintStmt::EntryType> &entries,
+                                  T &&t,
+                                  Args &&... values) {
+    entries.push_back(EntryType{t});
+    make_entries_helper(entries, std::forward<Args>(values)...);
+  }
+
+  template <typename... Args>
+  static std::vector<EntryType> make_entries(Args &&... values) {
+    std::vector<EntryType> ret;
+    make_entries_helper(ret, std::forward<Args>(values)...);
+    return ret;
+  }
+};
+
+class ConstStmt : public Stmt {
+ public:
+  LaneAttribute<TypedConstant> val;
+
+  ConstStmt(const LaneAttribute<TypedConstant> &val) : val(val) {
+    width() = val.size();
+    element_type() = val[0].dt;
+    for (int i = 0; i < ret_type.width; i++) {
+      TI_ASSERT(val[0].dt == val[i].dt);
+    }
+    TI_STMT_REG_FIELDS;
+  }
+
+  void repeat(int factor) override {
+    Stmt::repeat(factor);
+    val.repeat(factor);
+  }
+
+  bool has_global_side_effect() const override {
+    return false;
+  }
+
+  std::unique_ptr<ConstStmt> copy();
+
+  TI_STMT_DEF_FIELDS(ret_type, val);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+// General range for
+class RangeForStmt : public Stmt {
+ public:
+  Stmt *begin, *end;
+  std::unique_ptr<Block> body;
+  bool reversed;
+  int vectorize;
+  int parallelize;
+  int block_dim;
+  bool strictly_serialized;
+
+  RangeForStmt(Stmt *begin,
+               Stmt *end,
+               std::unique_ptr<Block> &&body,
+               int vectorize,
+               int parallelize,
+               int block_dim,
+               bool strictly_serialized);
+
+  bool is_container_statement() const override {
+    return true;
+  }
+
+  void reverse() {
+    reversed = !reversed;
+  }
+
+  std::unique_ptr<Stmt> clone() const override;
+
+  TI_STMT_DEF_FIELDS(begin,
+                     end,
+                     reversed,
+                     vectorize,
+                     parallelize,
+                     block_dim,
+                     strictly_serialized);
+  TI_DEFINE_ACCEPT
+};
+
+// for stmt over a structural node
+class StructForStmt : public Stmt {
+ public:
+  SNode *snode;
+  std::unique_ptr<Block> body;
+  std::unique_ptr<Block> block_initialization;
+  std::unique_ptr<Block> block_finalization;
+  std::vector<int> index_offsets;
+  int vectorize;
+  int parallelize;
+  int block_dim;
+  ScratchPadOptions scratch_opt;
+
+  StructForStmt(SNode *snode,
+                std::unique_ptr<Block> &&body,
+                int vectorize,
+                int parallelize,
+                int block_dim);
+
+  bool is_container_statement() const override {
+    return true;
+  }
+
+  std::unique_ptr<Stmt> clone() const override;
+
+  TI_STMT_DEF_FIELDS(snode,
+                     index_offsets,
+                     vectorize,
+                     parallelize,
+                     block_dim,
+                     scratch_opt);
+  TI_DEFINE_ACCEPT
+};
+
+class FuncBodyStmt : public Stmt {
+ public:
+  std::string funcid;
+  std::unique_ptr<Block> body;
+
+  FuncBodyStmt(const std::string &funcid, std::unique_ptr<Block> &&body);
+
+  bool is_container_statement() const override {
+    return true;
+  }
+
+  std::unique_ptr<Stmt> clone() const override;
+
+  TI_STMT_DEF_FIELDS(funcid);
+  TI_DEFINE_ACCEPT
+};
+
+class FuncCallStmt : public Stmt {
+ public:
+  std::string funcid;
+
+  FuncCallStmt(const std::string &funcid) : funcid(funcid) {
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(ret_type, funcid);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class KernelReturnStmt : public Stmt {
+ public:
+  Stmt *value;
+
+  KernelReturnStmt(Stmt *value, DataType dt) : value(value) {
+    this->ret_type = VectorType(1, dt);
+    TI_STMT_REG_FIELDS;
+  }
+
+  TI_STMT_DEF_FIELDS(value);
+  TI_DEFINE_ACCEPT_AND_CLONE
+};
+
+class WhileStmt : public Stmt {
+ public:
+  Stmt *mask;
+  std::unique_ptr<Block> body;
+
+  WhileStmt(std::unique_ptr<Block> &&body);
+
+  bool is_container_statement() const override {
+    return true;
+  }
+
+  std::unique_ptr<Stmt> clone() const override;
+
+  TI_STMT_DEF_FIELDS(mask);
+  TI_DEFINE_ACCEPT
+};
+
 class PragmaSLPStmt : public Stmt {
  public:
   int slp_width;
